requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Google Cloud APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0

# AI Services
openai==1.3.0

# HTTP Client
requests==2.31.0
httpx==0.25.2

# File handling
python-multipart==0.0.6

# Configuration
python-dotenv==1.0.0

# Logging & Monitoring (opcional para Cloud Run)
google-cloud-logging==3.8.0

# Security
cryptography==41.0.7